backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
sentence-transformers>=2.7.0
chromadb>=0.4.0
google-generativeai>=0.5.0
pandas>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
httpx>=0.24.0
lxml>=4.9.0
